requirements.txt
requests>=2.25.0
schedule>=1.1.0
PyYAML>=6.0.1 